+++ lib/oneview-sdk/resource/api200/enclosure_group.rb
@@ -49,6 +49,7 @@ module OneviewSDK
 
       # Adds the logical interconnect group
       # @param [OneviewSDK::LogicalInterconnectGroup] lig Logical Interconnect Group
+      # @return [OneviewSDK::EnclosureGroup] self
       def add_logical_interconnect_group(lig)
         lig.retrieve! unless lig['uri']
         lig['interconnectMapTemplate']['interconnectMapEntryTemplates'].each do |entry|
@@ -56,9 +57,11 @@ module OneviewSDK
             add_lig_to_bay(location['relativeValue'], lig) if location['type'] == 'Bay' && entry['permittedInterconnectTypeUri']
           end
         end
+        self
       end
 
       # Creates the interconnect bay mapping
+      # @return [OneviewSDK::EnclosureGroup] self
       def create_interconnect_bay_mapping
         @data['interconnectBayMappings'] = []
         1.upto(@data['interconnectBayMappingCount']) do |bay_number|
@@ -68,6 +71,7 @@ module OneviewSDK
           }
           @data['interconnectBayMappings'] << entry
         end
+        self
       end
 
       private